Thu, Jan 20, 2022, 11:56 AM +00:00

+++ credit/test.c
@@ -1,20 +1,20 @@
 #include <cs50.h>
 #include <stdio.h>
 
-int triangle(int a, int b, int c);
+float triangle(float a, float b, float c);
 
 int main(void)
 {
-	int x = get_int("One of the side: ");
-	int y = get_int("One of the side: ");
-	int z = get_int("One of the side: ");
+	float x = get_int("One of the side: ");
+	float y = get_int("One of the side: ");
+	float z = get_int("One of the side: ");
 
-	
+
 
 }
 
 
-int triangle(int x, int y, int z)
+float triangle(float x, float y, float z)
 {
 	if(a + b > c || a + c > b || c + b > a)
 	{
